// frontend/src/app/api/models/api-home-blog-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface apiHomeBlogModel {
  action_link: (string | null);
  action_text: (string | null);
  affiliate_id: number;
  content: string;
  created_at: string;
  id: string;
  image_link: (string | null);
  subtitle: (string | null);
  title: string;
  year: number;
}